requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.0
grpcio==1.59.0
h5py==3.10.0
hugchat==0.2.9
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
langchain==0.0.314
langsmith==0.0.43
libclang==16.0.6
llama-index==0.8.45.post1
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pypdf==3.16.4
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
streamlit==1.27.1
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.17
validators==0.22.0
Werkzeug==3.0.0
wrapt==1.14.1
yarl==1.9.2
zipp==3.17.0
